Finish moving stuff around
