+++ registry/api/v2/upload_sessions_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/gofrs/uuid/v5"
+	"github.com/robinkb/cascade-registry/registry/repository"
 	"github.com/robinkb/cascade-registry/registry/store"
-	"github.com/robinkb/cascade-registry/repository"
 	. "github.com/robinkb/cascade-registry/testing"
 	testclient "github.com/robinkb/cascade-registry/testing/client"
 	"github.com/robinkb/cascade-registry/testing/mock"
