+++ registry/repository/manifests_test.go
