+++ registry/repository/service_test.go
