+++ registry/repository/uploads_test.go
