+++ registry/repository/validate.go
