+++ registry/repository/validate_test.go
